move Degree.kt to kohomology/dg/degree/
