+++ kohomology/src/commonMain/kotlin/com/github/shwaka/kohomology/dg/DGAlgebra.kt
@@ -1,5 +1,6 @@
 package com.github.shwaka.kohomology.dg
 
+import com.github.shwaka.kohomology.dg.degree.Degree
 import com.github.shwaka.kohomology.linalg.Matrix
 import com.github.shwaka.kohomology.linalg.MatrixSpace
 import com.github.shwaka.kohomology.linalg.NumVector
