+++ kohomology/src/commonMain/kotlin/com/github/shwaka/kohomology/dg/GAlgebra.kt
@@ -1,6 +1,8 @@
 package com.github.shwaka.kohomology.dg
 
 import com.github.h0tk3y.betterParse.grammar.parseToEnd
+import com.github.shwaka.kohomology.dg.degree.Degree
+import com.github.shwaka.kohomology.dg.degree.DegreeMonoid
 import com.github.shwaka.kohomology.linalg.Matrix
 import com.github.shwaka.kohomology.linalg.MatrixSpace
 import com.github.shwaka.kohomology.linalg.NumVector
