+++ kohomology/src/commonMain/kotlin/com/github/shwaka/kohomology/dg/GLinearMap.kt
@@ -1,5 +1,6 @@
 package com.github.shwaka.kohomology.dg
 
+import com.github.shwaka.kohomology.dg.degree.Degree
 import com.github.shwaka.kohomology.exception.IllegalContextException
 import com.github.shwaka.kohomology.exception.InvalidSizeException
 import com.github.shwaka.kohomology.linalg.Matrix
