+++ kohomology/src/commonMain/kotlin/com/github/shwaka/kohomology/model/FreeLoopSpace.kt
@@ -3,7 +3,7 @@ package com.github.shwaka.kohomology.model
 import com.github.shwaka.kohomology.dg.DGAlgebraMap
 import com.github.shwaka.kohomology.dg.Derivation
 import com.github.shwaka.kohomology.dg.GAlgebraMap
-import com.github.shwaka.kohomology.dg.IntDegree
+import com.github.shwaka.kohomology.dg.degree.IntDegree
 import com.github.shwaka.kohomology.free.FreeDGAlgebra
 import com.github.shwaka.kohomology.free.FreeGAlgebra
 import com.github.shwaka.kohomology.free.IndeterminateName
